+++ Thunderstore/CHANGELOG.md
@@ -1,8 +1,14 @@
+## 1.0.3
+* Added a plushie.
+* Gave the vending machine a higher chance of spawning earlier on in the day.
+* Fixed bug that causes the EnableEnemyMesh function to break and then in turn break the vending machine.
+* Fixed bug where you could see two scan nodes on the company cola.
+
 ## 1.0.2
-* Fixed bug to do with the Vending Machine Registry not being able to handle null string ids
+* Fixed bug to do with the Vending Machine Registry not being able to handle null string ids.
 
 ## 1.0.1
-* Added cached spawn values for Experimentation, Assurance, Vow and Titan
+* Added cached spawn values for Experimentation, Assurance, Vow and Titan.
 
 ## 1.0.0
-* Initial Release
+* Initial Release.